[v0.9.8.1-qfit3-resol1-cln3-newf32-dbg2-sz2] more clean up

+++ indiv.h
@@ -33,9 +33,9 @@
 extern void
 rate_super_players 	( bool_t quiet
 					, struct ENC *enc
-					, long N_enc
+					, size_t N_enc
 					, int *performance_type
-					, int n_players
+					, size_t n_players
 					, double *ratingof
 					, double white_advantage
 					, bool_t *flagged
